remove notes from ARC antiracism forumn
In the meeting of ARC tonight, the decision was taken to in committee to discontinue open access to the “Notes” and  recording from the Forum on Anti-Racism that has been previously published.

The Adult Religious Education Committee has determined that this action need be taken to protect the confidentiality of the commenters, whose faces and names may appear in the recording and notes.

Accordingly, please take direct action now and confirm back to Co-Clerks, Astra and Tom whose emails are in the header that this is done, and you may address any questions to them.

Many Thanks ~ PM